requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
DateTime==4.7
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
pytz==2022.6
requests==2.28.1
urllib3==1.26.13
zope.interface==5.5.2
